+++ tests/viz_test.py
@@ -1,0 +1,15 @@
+import matplotlib
+import matplotlib.pyplot as plt
+
+def test_histogram(df_trimmed):
+    df = df_trimmed
+    matplotlib.use('agg')
+    plt.figure()
+    df.x.viz.histogram(show=False)
+
+
+def test_heatmap(df_trimmed):
+    df = df_trimmed
+    matplotlib.use('agg')
+    plt.figure()
+    df.x.viz.histogram(show=False)